MER-201 Moved access denied handler
