+++ vulcanus/src/main/java/com/merkury/vulcanus/security/CustomAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.merkury.vulcanus.exception;
+package com.merkury.vulcanus.security;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
